requirements.txt
fitz==0.0.1.dev2
#Pillow==10.0.1
Pillow==10.3.0
PyPDF2==3.0.1
Requests==2.32.2
streamlit==1.34.0
